refactor: section 사용 안하는 방식
Create Button의 Modal에 값이 필요하기 때문
