+++ src/containers/CommunityPostCreateButton/CommunityPostCreateButton.tsx
@@ -1,0 +1,2 @@
+function CommunityPostCreateButton() {}
+export default CommunityPostCreateButton;
